no poder eliminar iglesia si tiene ministerios asociados

+++ src/services/iglesiasService.js
@@ -1,4 +1,5 @@
 const Iglesia = require('../models/iglesia');
+const Ministerio = require('../models/ministerio');
 
 exports.getAllIglesias = async () => {
   return await Iglesia.find();
@@ -18,6 +19,14 @@ exports.updateIglesia = async (id, data) => {
 };
 
 exports.deleteIglesia = async (id) => {
+  // Verificar dependencias: ministerios asociados a esta iglesia
+  const ministeriosAsociados = await Ministerio.countDocuments({ id_iglesia: id });
+  if (ministeriosAsociados > 0) {
+    const err = new Error(`No se puede eliminar la iglesia porque existen ${ministeriosAsociados} ministerio(s) asociados.`);
+    err.statusCode = 400; // Regla de negocio
+    throw err;
+  }
+
   const result = await Iglesia.findByIdAndDelete(id);
   return !!result;
 };
